requirements.txt
numpy==2.4.2
opencv-python==4.13.0.92
pillow==12.1.1
python-dotenv==1.2.1
pillow-heif
